fix table realtions issues

+++ db/migrate/20221023135953_create_locations.rb
@@ -2,8 +2,8 @@ class CreateLocations < ActiveRecord::Migration[7.0]
   def change
     create_table :locations do |t|
       t.string :name
-      t.decimal :longitude
-      t.decimal :latitude
+      t.integer :longitude
+      t.integer :latitude
 
       t.timestamps
     end
